move core => exports/raw
