+++ exports/buffer.test.ts
@@ -1,16 +1,16 @@
 import assert from "node:assert";
 import { buffer } from "node:stream/consumers";
 import { describe, it, mock } from "node:test";
-import { CompressionMethod } from "../core/compression-core.ts";
-import { ZipPlatform, ZipVersion } from "../core/constants.ts";
-import { UnixFileAttributes } from "../core/file-attributes.ts";
-import { ZipEntry } from "../core/zip-entry.ts";
 import {
   EmptyZip32,
   Zip32WithThreeEntries,
   generateZip,
 } from "../test-util/fixtures.ts";
 import { ZipBufferReader } from "./buffer.ts";
+import { CompressionMethod } from "./raw/compression-core.ts";
+import { ZipPlatform, ZipVersion } from "./raw/constants.ts";
+import { UnixFileAttributes } from "./raw/file-attributes.ts";
+import { ZipEntry } from "./zip-entry.ts";
 
 describe("web/buffer", () => {
   describe("ZipBufferReader", () => {
