+++ exports/compression.test.ts
@@ -1,10 +1,10 @@
 import assert from "node:assert";
 import { buffer, text } from "node:stream/consumers";
 import { describe, it } from "node:test";
-import { CompressionMethod } from "../core/compression-core.ts";
 import { asyncIterable, base64, base64iterable } from "../test-util/data.ts";
 import { maxChunkSize } from "../util/streams.ts";
 import { defaultCompressors, defaultDecompressors } from "./compression.ts";
+import { CompressionMethod } from "./raw/compression-core.ts";
 
 describe("web/compression", () => {
   describe("defaultCompressors", () => {
