+++ exports/raw/zip-trailer.test.ts
@@ -1,6 +1,6 @@
 import assert from "node:assert";
 import { describe, it } from "node:test";
-import { assertBufferEqual } from "../test-util/assert.ts";
+import { assertBufferEqual } from "../../test-util/assert.ts";
 import {
   bigUint,
   cp437,
@@ -9,9 +9,13 @@ import {
   longUint,
   shortUint,
   tinyUint,
-} from "../test-util/data.ts";
+} from "../../test-util/data.ts";
+import {
+  MultiDiskError,
+  ZipFormatError,
+  ZipSignatureError,
+} from "../errors.ts";
 import { ZipPlatform, ZipVersion } from "./constants.ts";
-import { MultiDiskError, ZipFormatError, ZipSignatureError } from "./errors.ts";
 import { Eocdr, Zip64Eocdl, Zip64Eocdr, ZipTrailer } from "./zip-trailer.ts";
 
 describe("core/zip-trailer", () => {
