+++ exports/reader.ts
@@ -1,16 +1,3 @@
-import { CentralDirectoryHeader } from "../core/central-directory-header.ts";
-import {
-  decompress,
-  type CompressionAlgorithms,
-} from "../core/compression-core.ts";
-import { LocalFileHeader } from "../core/local-file-header.ts";
-import { ZipEntry } from "../core/zip-entry.ts";
-import {
-  Eocdr,
-  Zip64Eocdl,
-  Zip64Eocdr,
-  ZipTrailer,
-} from "../core/zip-trailer.ts";
 import { assert } from "../util/assert.ts";
 import { asyncDisposeOrClose } from "../util/disposable.ts";
 import { lazy } from "../util/lazy.ts";
@@ -20,6 +7,19 @@ import {
 } from "../util/streams.ts";
 import type { Constructor } from "../util/type-utils.ts";
 import { defaultDecompressors } from "./compression.ts";
+import { CentralDirectoryHeader } from "./raw/central-directory-header.ts";
+import {
+  decompress,
+  type CompressionAlgorithms,
+} from "./raw/compression-core.ts";
+import { LocalFileHeader } from "./raw/local-file-header.ts";
+import {
+  Eocdr,
+  Zip64Eocdl,
+  Zip64Eocdr,
+  ZipTrailer,
+} from "./raw/zip-trailer.ts";
+import { ZipEntry } from "./zip-entry.ts";
 
 const DefaultBufferSize = 1024 ** 2;
